for #2556, move package of AbstractBaseIntegrateSQLParsingTest
