+++ sharding-core/sharding-core-parse/sharding-core-parse-test/src/test/java/org/apache/shardingsphere/core/parse/integrate/engine/AbstractBaseIntegrateSQLParsingTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.parse.integrate.engine.sharding;
+package org.apache.shardingsphere.core.parse.integrate.engine;
 
 import com.google.common.base.Preconditions;
 import lombok.AccessLevel;
